+++ test/fixture/coords.ts
@@ -1,4 +1,4 @@
-export const coordsA: GeolocationCoordinates = {
+export const coordsA: GeolocationCoordinatesWithToStringTag = {
   latitude: 40.71703581534977,
   longitude: -74.03457283319447,
   accuracy: 25.019,
@@ -6,9 +6,10 @@ export const coordsA: GeolocationCoordinates = {
   altitudeAccuracy: 9.838127136230469,
   heading: null,
   speed: null,
+  [Symbol.toStringTag]: "GeolocationCoordinates",
 };
 
-export const coordsB: GeolocationCoordinates = {
+export const coordsB: GeolocationCoordinatesWithToStringTag = {
   latitude: 12,
   longitude: 34,
   accuracy: 56,
@@ -16,9 +17,10 @@ export const coordsB: GeolocationCoordinates = {
   altitudeAccuracy: 9,
   heading: null,
   speed: null,
+  [Symbol.toStringTag]: "GeolocationCoordinates",
 };
 
-export const coordsC: GeolocationCoordinates = {
+export const coordsC: GeolocationCoordinatesWithToStringTag = {
   latitude: 98,
   longitude: 76,
   accuracy: 54,
@@ -26,9 +28,10 @@ export const coordsC: GeolocationCoordinates = {
   altitudeAccuracy: 10,
   heading: null,
   speed: null,
+  [Symbol.toStringTag]: "GeolocationCoordinates",
 };
 
-export const coordsD: GeolocationCoordinates = {
+export const coordsD: GeolocationCoordinatesWithToStringTag = {
   latitude: 23,
   longitude: 45,
   accuracy: 67,
@@ -36,4 +39,9 @@ export const coordsD: GeolocationCoordinates = {
   altitudeAccuracy: 10,
   heading: null,
   speed: null,
+  [Symbol.toStringTag]: "GeolocationCoordinates",
+};
+
+type GeolocationCoordinatesWithToStringTag = GeolocationCoordinates & {
+  [Symbol.toStringTag]: string;
 };